requirements.txt
pytest~=7.4.4
h11~=0.14.0
pip~=24.0
attrs~=22.2.0
typing_extensions~=4.9.0
trio~=0.24.0
outcome~=1.3.0.post0
sniffio~=1.3.0
sortedcontainers~=2.4.0
cffi~=1.16.0
idna~=3.6
Appium-Python-Client~=3.1.1
selenium~=4.17.0
urllib3~=2.1.0
pluggy~=1.3.0
iniconfig~=2.0.0
certifi~=2023.11.17
pycparser~=2.21
websocket-client~=0.59.0
six~=1.16.0
eyes-selenium~=5.24.0
marshmallow~=2.21.0
wsproto~=1.2.0
requests~=2.31.0
python-dotenv~=1.0.0
packaging~=23.2
webdriver-manager~=4.0.1
smart-assertions~=1.0.2
allure-pytest~=2.13.2

